+++ src/services/reducers/rootReducer.js
@@ -1,0 +1,5 @@
+import { combineReducers } from 'redux'
+
+export const rootReducer = combineReducers({
+
+})
