+++ src/utils/initialState.js
@@ -1,0 +1,6 @@
+export const initialState = {
+    allIngredients: [],
+    currentConstructorIngredients: [],
+    currentIngredient: null,
+    createdOrder: null
+}
